% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaSimHlpR.R
\name{initiatePop}
\alias{initiatePop}
\title{Title}
\usage{
initiatePop(
  nFounders,
  nChr,
  segSites,
  schemeParams,
  nQTL,
  nSNP,
  genVar,
  meanDD,
  varDD
)
}
\arguments{
\item{varDD}{}
}
\value{

}
\description{
Title
}
