% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaSimHlpR.R
\name{runPhenoTrial}
\alias{runPhenoTrial}
\title{runPhenoTrial function}
\usage{
runPhenoTrial(
  nReps,
  nLocs,
  nChecks,
  entries,
  checks = NULL,
  SP,
  errVars,
  trialName,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
runPhenoTrial function
}
