% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaSimHlpR.R
\name{selectAndCross}
\alias{selectAndCross}
\title{Title}
\usage{
selectAndCross(
  candSelCrit,
  col = "BLUP",
  candPop,
  nParents,
  nProgeny,
  nCrosses,
  SP
)
}
\arguments{
\item{SP}{}
}
\value{

}
\description{
Title
}
